requirements.txt
requests==2.32.4
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3